+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- On `ChessGame`, added heavyweight constructor.
 - On `ChessGame.makeAllMove*` methods, added showing where the error move occurred. (e.g.
   `Illegal move detected! Move : Nc9, FEN : rnbqkbnr/... (Failed at move #4 in sequence "e4 e5 Nf3 Nc9")`)
 - On `ChessGame` methods throwing `MoveNotFoundException`, added node id on exception message.